py-src/requirements.txt
matplotlib
hdrhistogram
pandas
numpy